LW1/LW1.6 updated

+++ LW1/LW1.6/include/opt_handler.h
@@ -1,3 +1,3 @@
 #pragma once
 
-double get_precision(const int argc,const char** argv);
+long double get_precision(const int argc,const char** argv);
